영화 입력 백엔드 구현
트랜잭션 처리
film 테이블 입력,
film_category 테이블 입력
filmMapper, filmService 구현 완료

입력폼 구현 X
컨트롤러 구현 X

+++ sakila/src/main/java/com/gd/sakila/vo/Film.java
@@ -11,9 +11,9 @@ public class Film {
 	private int languageId;
 	private int originalLanguageId;
 	private int rentalDuration;
-	private int rentalRate;
+	private double rentalRate;
 	private int length;
-	private int replacementCost;
+	private double replacementCost;
 	private String rating;
 	private String specialFeatures;
 	private String lastUpdate;
